Update fsr.f90
User defined type FastSonicData abolished, and replaced with its components, as an attempt to solve issue #8. Now all works.

I've no absolute prove of this, but the OO version of code seemed to be syntactically and semantically correct. If so, its failure to operate (and hence #8) is a compiler bug. In conclusion: it's better to consider the PGI Fortran as if it is not Fortran 2003 compliant, whatever its documentation states.